Feat

+++ web/app/lib/services/service.service.ts
@@ -1,19 +1,43 @@
 import { prisma } from "../prisma";
+import { put, del } from "@vercel/blob";
 import { CreateService, EditService } from "../utils/requests/service.request";
 import { ServiceResponses } from "../utils/responses/serviceResponses.";
 import { AppError } from "../withError";
+import { KitType } from "@prisma/client";
+
+type FileOrString = File | string;
 
 export type ServiceSearchPayload = {
   query: string;
 };
 
 export const ServiceService = {
-  async create(content: CreateService) {
+  async create(formData: FormData) {
+    const { name, icon, price, forKit }: CreateService = {
+      name: String(formData.get("name")),
+      price: Number(formData.get("price")),
+      icon: formData.get("icon") as File,
+      forKit: formData.get("forKit") as KitType
+    }
+
+    const serviceId = crypto.randomUUID();
+
+    const blob = await put(
+      `services/${serviceId}/${crypto.randomUUID()}`,
+      icon,
+      { access: "public" }
+    );
+
+    const forKitType = Object.values(KitType).includes(forKit) ? forKit : "ALL";
+
     try {
       return await prisma.service.create({
         data: {
-          name: content.name.trim().normalize("NFC").toLowerCase(),
-          price: content.price
+          id: serviceId,
+          name: name.trim().normalize("NFC").toLowerCase(),
+          price: price,
+          icon: blob.url,
+          forKit: forKitType
         }
       });
 
@@ -46,11 +70,35 @@ export const ServiceService = {
     return services;
   },
 
-  async getAll() {
-    return await prisma.service.findMany();
+  async getAll(kitType: KitType) {
+    const param = Object.values(KitType).includes(kitType) ? kitType : "ALL";
+    const filter = [param];
+
+    if (filter.includes("ALL")) {
+      return await prisma.service.findMany();
+    }
+
+    filter.push("ALL");
+
+    return await prisma.service.findMany({
+      where: {
+        forKit: {
+          in: filter
+        }
+      }
+    });
   },
 
-  async edit(id: string, content: EditService) {
+  async edit(id: string, formData: FormData) {
+    const { name, icon, price, forKit }: EditService = {
+      name: String(formData.get("name")),
+      price: Number(formData.get("price")),
+      icon: formData.get("icon") as FileOrString,
+      forKit: formData.get("forKit") as KitType
+    }
+
+    const forKitType = Object.values(KitType).includes(forKit) ? forKit : "ALL";
+
     try {
       const currentService = await prisma.service.findUnique({
         where: {
@@ -62,17 +110,35 @@ export const ServiceService = {
 
       let serviceEdited: boolean = false;
 
+      let image: string;
+
+      if (icon instanceof File) {
+        image = await put(
+          `services/${id}/${crypto.randomUUID()}`,
+          icon,
+          { access: "public" }
+        ).then(blob => blob.url);
+
+        await del(currentService.icon);
+
+      } else {
+        image = icon;
+      }
+
       if (
-        currentService.name !== content.name ||
-        !currentService.price.equals(content.price)
+        currentService.name !== name ||
+        !currentService.price.equals(price) ||
+        currentService.forKit !== forKit
       ) {
         await prisma.service.update({
           where: {
             id
           },
           data: {
-            name: content.name.trim().normalize("NFC").toLowerCase(),
-            price: content.price
+            name: name.trim().normalize("NFC").toLowerCase(),
+            price: price,
+            icon: image,
+            forKit: forKitType
           }
         });
 
@@ -93,6 +159,14 @@ export const ServiceService = {
 
   async delete(id: string) {
     try {
+      const service = await prisma.service.findUnique({
+        where: { id }
+      });
+
+      if (!service) throw new AppError(404, ServiceResponses.SERVICE_NOT_FOUND);
+
+      await del(service.icon);
+
       await prisma.service.delete({
         where: {
           id
@@ -100,7 +174,7 @@ export const ServiceService = {
       });
 
     } catch {
-      throw new AppError(404, ServiceResponses.SERVICE_DELETED_ERROR);
+      throw new AppError(400, ServiceResponses.SERVICE_DELETED_ERROR);
     }
   },
